Remove hardcoded image name
Remove hardcoded image name, to make it easier to use the same/similar script more than once (IP addresses might also have to be changed, did that manually for now). Could also move some of this to input params for even better practical use, but agreed that would be outside the scope of this starting script.